refactor(BookmarkButton): extract bookmark button into separate component for reusability
